Update
- Cleaned up configs.
- Added Docker env.
- Added PHPStorm configs.
- Added CI build instructions.
- Removed unused services.
