+++ .env.example
@@ -1,0 +1,8 @@
+BASE_PATH=./
+BUILD_ROOT_PATH=/app/
+PROJECT_NAME=dhii_delimited-token-template
+
+PHP_BUILD_VERSION=7.1
+PHP_TEST_VERSION=7.1
+
+HOST_IP_ADDRESS=127.0.0.1
