+++ .gitattributes
@@ -7,3 +7,5 @@ nbproject export-ignore
 .gitignore export-ignore
 composer.lock export-ignore
 phpunit.xml export-ignore
+
+* text eol=lf
