+++ .gitignore
@@ -1,5 +1,5 @@
 /vendor/
 /nbproject/private/
-/test/coverage/
-/test/log/
+/tests/coverage/
+/tests/log/
 /build/
