+++ .idea/.gitignore
@@ -1,0 +1,7 @@
+# Default ignored files
+/shelf/
+/workspace.xml
+# Datasource local storage ignored files
+/dataSources.local.xml
+# Editor-based HTTP Client requests
+/httpRequests/
