+++ composer.json
@@ -12,14 +12,13 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^5.3 | ^7.0"
+        "php": "^7.1 | ^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "ptrofimov/xpmock": "^1.1",
-        "dhii/php-cs-fixer-config": "dev-php-5.3",
-        "codeclimate/php-test-reporter": "<=0.3.2",
-        "dhii/stringable-interface": "^0.1.0"
+        "phpunit/phpunit": "^6.0 | ^7.0 | ^8.0 | ^9.0",
+        "vimeo/psalm": "^3.11.7 | ^4.0",
+        "slevomat/coding-standard": "^6.0",
+        "symfony/polyfill-php80": "^1.19"
     },
     "suggest": {
         "dhii/i18n-interface": "For internationalizing and translating human readable strings."
@@ -27,6 +26,11 @@
     "autoload": {
         "psr-4": {
             "Dhii\\Util\\String\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Dhii\\Util\\String\\FuncTest": "tests/functional"
         }
     },
     "scripts": {
