--- test/functional/DescriptionAwareInterfaceTest.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace Dhii\Util\String\FuncTest;
-
-use Dhii\Util\String\DescriptionAwareInterface;
-use Xpmock\TestCase;
-
-/**
- * Tests {@see Dhii\Util\String\DescriptionAwareInterface}.
- *
- * @since [*next-version*]
- */
-class DescriptionAwareInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Util\String\DescriptionAwareInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return DescriptionAwareInterface
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-                     ->getDescription();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject,
-            'Subject is not a valid instance'
-        );
-    }
-}
