--- test/functional/LabelAwareInterfaceTest.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace Dhii\Util\String\FuncTest;
-
-use Dhii\Util\String\LabelAwareInterface;
-use Xpmock\TestCase;
-
-/**
- * Tests {@see Dhii\Util\String\LabelAwareInterface}.
- *
- * @since [*next-version*]
- */
-class LabelAwareInterfaceTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Util\String\LabelAwareInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @return LabelAwareInterface
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-                     ->getLabel();
-
-        return $mock->new();
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject,
-            'Subject is not a valid instance'
-        );
-    }
-}
